Parallel EG EA &Solve Preimage Error (#268)
resolve Fail at 157 - (1) preimage already integrated by delete double broadcasting and extra checking in processPreimage
make EG EA be included in block in the same time (Parallel)
Make Block EG process Update
solve basePort problem

+++ grandpa/play_round_test.go
@@ -126,7 +126,7 @@ func TestPlayRound(t *testing.T) {
 		node.PlayGrandpaRound(ctx, 1)
 	}
 	tmp_blk := genesis_blk
-	basePort := 10000
+	basePort := uint16(10000)
 	graph_server := types.NewGraphServer(basePort)
 	go graph_server.StartServer()
 	ticker := time.NewTicker(6 * time.Second)
